feat(mass-profiles): add multiple new lensing mass profile implementations (migrate from lenstronomy)
Add TNFW, multipole, flexion, dipole, and pseudo-Jaffe mass profiles to extend
the library's gravitational lensing modeling capabilities. The new profiles
include both spherical and elliptical variants, with comprehensive test suites
validating against lenstronomy reference implementations.

- Add TNFW spherical and elliptical potential profiles with truncation support
- Implement circular and elliptical multipole expansions (m=1,3,4)
- Create composite EPL profiles with multipole terms (M3M4, M1M3M4, boxy/disky)
- Add flexion (cartesian and FG basis) and dipole response profiles
- Implement pseudo-Jaffe profiles (spherical and elliptical potential)
- Include comprehensive test suites with lenstronomy comparison tests
- Fix dtype consistency in lens mapping matrix construction

+++ TinyLensGpu/utils/lensing/mapping.py
@@ -74,7 +74,7 @@ def lens_mapping_matrix_from(
     col_indices = indices.flatten()
     weights_flat = weights.flatten()
     
-    map_mat = jnp.zeros((N_data, N_source), dtype=jnp.float32)
+    map_mat = jnp.zeros((N_data, N_source), dtype=weights_flat.dtype)
     map_mat = map_mat.at[row_indices, col_indices].set(weights_flat)
 
     return map_mat
